+++ src/Sentry/Internal/Constants.cs
@@ -33,11 +33,5 @@ internal static class Constants
     public const string DebugEnvironmentSetting = "debug";
 
     // See: https://github.com/getsentry/sentry-release-registry
-#if ANDROID
-        public const string SdkName = "sentry.dotnet.android";
-#elif __IOS__
-    public const string SdkName = "sentry.dotnet.cocoa";
-#else
-        public const string SdkName = "sentry.dotnet";
-#endif
+    public const string SdkName = "sentry.dotnet";
 }